Convert MachineLearningFlyout to hooks
